More stuff
mostly fixes
