+++ src/main/java/net/dv8tion/jda/bot/entities/ApplicationInfo.java
@@ -20,7 +20,6 @@ import net.dv8tion.jda.core.JDA;
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.ISnowflake;
 import net.dv8tion.jda.core.entities.User;
-import net.dv8tion.jda.core.utils.MiscUtil;
 import java.util.Arrays;
 import java.util.Collection;
