+++ src/main/java/net/dv8tion/jda/bot/entities/impl/ApplicationInfoImpl.java
@@ -21,7 +21,6 @@ import net.dv8tion.jda.core.JDA;
 import net.dv8tion.jda.core.Permission;
 import net.dv8tion.jda.core.entities.User;
 
-import java.util.Arrays;
 import java.util.Collection;
 
 public class ApplicationInfoImpl implements ApplicationInfo
