+++ src/main/java/net/dv8tion/jda/client/events/relationship/UserBlockedEvent.java
@@ -29,6 +29,6 @@ public class UserBlockedEvent extends GenericRelationshipAddEvent
 
     public BlockedUser getBlockedUser()
     {
-        return (BlockedUser) getBlockedUser();
+        return (BlockedUser) getRelationship();
     }
 }
