+++ src/main/java/net/dv8tion/jda/core/utils/Checks.java
@@ -73,10 +73,25 @@ public class Checks
             throw new IllegalArgumentException(name + " may not be empty");
     }
 
+    public static void notEmpty(final Object[] argument, final String name)
+    {
+        notNull(argument, name);
+        if (argument.length == 0)
+            throw new IllegalArgumentException(name + " may not be empty");
+    }
+
     public static void noneNull(final Collection<?> argument, final String name)
     {
         notNull(argument, name);
         argument.forEach(it -> notNull(it, name));
+    }
+
+    public static void noneNull(final Object[] argument, final String name)
+    {
+        notNull(argument, name);
+        for (Object it : argument) {
+            notNull(it, name);
+        }
     }
 
     public static <T extends CharSequence> void noneEmpty(final Collection<T> argument, final String name)
